+++ exasol_script_languages_developer_sandbox/cli/commands/setup_ci_codebuild.py
@@ -3,7 +3,7 @@ from exasol_script_languages_developer_sandbox.cli.common import add_options
 from exasol_script_languages_developer_sandbox.cli.options.aws_options import aws_options
 from exasol_script_languages_developer_sandbox.cli.options.logging import logging_options
 from exasol_script_languages_developer_sandbox.lib.aws_access.aws_access import AwsAccess
-from exasol_script_languages_developer_sandbox.lib.ci_codebuild.ci_codebuild import run_setup_ci_codebuild
+from exasol_script_languages_developer_sandbox.lib.setup_ci_codebuild.ci_codebuild import run_setup_ci_codebuild
 from exasol_script_languages_developer_sandbox.lib.logging import set_log_level
