+++ setup.py
@@ -10,7 +10,7 @@ packages = \
  'exasol_script_languages_developer_sandbox.lib.ansible',
  'exasol_script_languages_developer_sandbox.lib.asset_printing',
  'exasol_script_languages_developer_sandbox.lib.aws_access',
- 'exasol_script_languages_developer_sandbox.lib.ci_codebuild',
+ 'exasol_script_languages_developer_sandbox.lib.setup_ci_codebuild',
  'exasol_script_languages_developer_sandbox.lib.export_vm',
  'exasol_script_languages_developer_sandbox.lib.setup_ec2',
  'exasol_script_languages_developer_sandbox.lib.vm_bucket',
